+++ monitor/services/pom.xml
@@ -193,23 +193,5 @@
 			<artifactId>javax.el-api</artifactId>
 			<scope>provided</scope>
 		</dependency>
-
-        <!-- REST documentation -->
-        <dependency>
-            <groupId>org.jvnet.ws.wadl</groupId>
-            <artifactId>wadl-core</artifactId>
-            <version>1.1.6</version>
-        </dependency>
-        <dependency>
-            <groupId>org.jvnet.ws.wadl</groupId>
-            <artifactId>wadl-client-plugin</artifactId>
-            <version>1.1.6</version>
-        </dependency>
-        <dependency>
-            <groupId>com.mangofactory</groupId>
-            <artifactId>swagger-springmvc</artifactId>
-            <version>1.0.2</version>
-        </dependency>
-
     </dependencies>
 </project>
